Breakout granule layer extend handling
